refactor: `SeparatorVisualTransformation`으로 구분자 변환 일반화
